+++ src/pages/options/Config.svelte
@@ -1,21 +1,42 @@
 <script lang="ts">
+	import Button from 'flowbite-svelte/Button.svelte';
 	import FloatingLabelInput from 'flowbite-svelte/FloatingLabelInput.svelte';
 	import Heading from 'flowbite-svelte/Heading.svelte';
 	import P from 'flowbite-svelte/P.svelte';
+	import { get } from 'svelte/store';
 	import Eye from '~/components/Eye.svelte';
+	import * as stores from '~/core/stores';
 	import { accessToken, clientID, clientSecret, refreshToken } from '~/core/stores';
+	import raindrop from '~/lib/raindrop';
 
 	let showClientID = false;
 	let showClientSecret = false;
 	let showAccessToken = false;
 	let showRefreshToken = false;
+
+	/**
+	 * Launch web auth flow.
+	 */
+	async function launchWebAuthFlow() {
+		const { accessToken, refreshToken } = await raindrop.auth.launchWebAuthFlow({
+			clientID: get(stores.clientID),
+			clientSecret: get(stores.clientSecret)
+		});
+		stores.accessToken.set(accessToken);
+		stores.refreshToken.set(refreshToken);
+	}
 </script>
 
 <div>
-	<Heading data-testid="heading" class="mt-2" tag="h2" customSize="text-xl font-bold"
-		>General</Heading
-	>
+	<Heading data-testid="heading" class="mt-2" tag="h2" customSize="text-xl font-bold">
+		General
+	</Heading>
 	<P data-testid="description" class="mt-2">Extension global settings.</P>
+	<P>
+		<!-- TODO: More detailed guide registering application with pictures -->
+		To register your application, <b>Client ID</b> and <b>Client Secret</b> then click
+		<b>Register</b>.
+	</P>
 	<div class="grid grid-cols-1">
 		<div class="mt-6 grid grid-cols-1 grid-flow-row gap-x-4 gap-y-6">
 			<div class="flex">
@@ -90,7 +111,11 @@
 					bind:open={showRefreshToken}
 				/>
 			</div>
-			<!-- TODO: Most feature of popup should be here, e.g.) initiating OAuth, token validity check, force refresh token, etc. -->
+		</div>
+		<!-- TODO: Debug actions: check token validity, force refresh token, etc. -->
+		<div class="mt-3 flex">
+			<!-- TODO: Validate inputs (client ID and secret) -->
+			<Button outline on:click={launchWebAuthFlow}>Register</Button>
 		</div>
 	</div>
 </div>
